requirements/base.txt
Django==1.7.1
smartmin==1.4.1
django-extensions==1.4.9
django-grappelli==2.6.3
django-phonenumber-field==0.6
distribute==0.7.3
MySQL-python==1.2.5
django-author==0.1
django-bootstrap3==5.0.3
django-celery==3.1.16
